refactor: Changed component names
Gave components a captial letter. Closes #202
