+++ src/lib/components/svg/PlayStory.svelte
